Fix some warnings in tests.
- Unused variables
- Signed/Unsigned comparisons.


cregit: 37320e36cd967a5823dc43fb87ce4342963e4ade https://github.com/qt/qtbase/commit/37320e36cd967a5823dc43fb87ce4342963e4ade

+++ tests/auto/widgets/kernel/qsizepolicy/tst_qsizepolicy.cpp
@@ -232,7 +232,7 @@ name|ITEMCOUNT
 parameter_list|(
 name|arr
 parameter_list|)
-value|sizeof(arr)/sizeof(arr[0])
+value|int(sizeof(arr)/sizeof(arr[0]))
 name|QSizePolicy
 name|sp
 decl_stmt|,
